+++ bank-feeds/RELEASES.md
@@ -819,3 +819,13 @@ Based on:
 - [python v9.0.0] bank-feeds
 ### Releases
 - [PyPI v9.0.0] https://pypi.org/project/codat-bankfeeds/9.0.0 - bank-feeds
+
+## 2024-12-10 10:08:17
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.453.8 (2.474.15) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v9.0.1] bank-feeds
+### Releases
+- [PyPI v9.0.1] https://pypi.org/project/codat-bankfeeds/9.0.1 - bank-feeds